LRQA-64470 Move files to bookmarks-item-selector-web
