--- modules/apps/bookmarks/bookmarks-web/src/main/resources/META-INF/resources/bookmarks/item/selector/init.jsp
@@ -1,25 +1,0 @@
-<%--
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
---%>
-
-<%@ include file="/init.jsp" %>
-
-<%@ page import="com.liferay.bookmarks.web.internal.display.context.BookmarkEntriesItemSelectorDisplayContext" %><%@
-page import="com.liferay.bookmarks.web.internal.display.context.BookmarkEntriesItemSelectorManagementToolbarDisplayContext" %><%@
-page import="com.liferay.item.selector.criteria.InfoItemItemSelectorReturnType" %><%@
-page import="com.liferay.portal.kernel.json.JSONObject" %><%@
-page import="com.liferay.portal.kernel.json.JSONUtil" %>
-
-<%@ include file="/bookmarks/init-ext.jsp" %>
